+++ modules/secrets-manager-secret/versions.tf
@@ -1,10 +1,10 @@
 terraform {
-  required_version = ">= 1.9"
+  required_version = ">= 1.11"
 
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 6.0, < 7.0"
+      version = ">= 6.50.0, < 7.0"
     }
     random = {
       source  = "hashicorp/random"